requirements.txt
pillow
requests
selenium
configparser
